+++ lib/cambiatus_web/views/email_view.ex
@@ -1,0 +1,3 @@
+defmodule CambiatusWeb.EmailView do
+  use CambiatusWeb, :view
+end
