+++ lib/cambiatus_web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule CambiatusWeb.LayoutView do
+  use CambiatusWeb, :view
+end
